refactor(engine): extract logic from GutterManager and MarkdownEditor
GutterManager.svelte (1693→1534 lines):
- gutter-manager.types.ts — interfaces (GutterItem, ProcessedAnchor, etc.)
- gutter-manager-utils.ts — pure functions (anchor processing, preview, cache)

MarkdownEditor.svelte (1084→924 lines):
- markdown-editor.types.ts — MarkdownEditorProps interface
- useImageUpload.svelte.ts — image upload composable (drag/drop/paste/upload)

Both components retain their templates and scoped styles (Svelte constraint)
but their logic surface is now independently readable.
